+++ packages/editor-core/test/browser/plugins/tasks-and-decisions/input-rules.ts
@@ -14,9 +14,18 @@ import {
 } from '../../../../src/test-helper';
 import tasksAndDecisionsPlugins from '../../../../src/plugins/tasks-and-decisions';
 import defaultSchema from '../../../../src/test-helper/schema';
+import uuid from '../../../../src/plugins/tasks-and-decisions/uuid';
 
 chai.use(chaiPlugin);
 describe('tasks and decisions - input rules', () => {
+
+  before(() => {
+    uuid.setStatic('local-decision');
+  });
+
+  after(() => {
+    uuid.setStatic(false);
+  });
 
   const editor = (doc: any) => makeEditor({
     doc,