+++ packages/emoji/src/api/EmojiUtils.ts
@@ -9,7 +9,6 @@ import {
   EmojiServiceDescriptionWithVariations,
   EmojiServiceRepresentation,
   EmojiServiceResponse,
-  ImageRepresentation,
   SpriteServiceRepresentation,
 } from '../types';
 import { isImageRepresentation, isSpriteServiceRepresentation } from '../type-helpers';
@@ -65,8 +64,9 @@ export const denormaliseServiceRepresentation = (representation: EmojiServiceRep
       };
     }
   } else if (isImageRepresentation(representation)) {
-    const { height, width, imagePath } = representation as ImageRepresentation;
+    const { height, width, imagePath } = representation;
     if (isMediaApiUrl(imagePath, meta)) {
+      // Convert to MediaRepresentation
       return {
         height,
         width,
@@ -104,7 +104,7 @@ export const denormaliseSkinEmoji = (skinEmojis?: EmojiServiceDescription[], met
  */
 export const denormaliseEmojiServiceResponse = (emojiData: EmojiServiceResponse): EmojiResponse  => {
   const emojis: EmojiDescription[] = emojiData.emojis.map((emoji: EmojiServiceDescriptionWithVariations): EmojiDescriptionWithVariations => {
-    const { id, name, shortName, type, category, order, fallback, ascii } = emoji;
+    const { id, name, shortName, type, category, order, fallback, ascii, searchable } = emoji;
     const representation = denormaliseServiceRepresentation(emoji.representation, emojiData.meta);
     const skinVariations = denormaliseSkinEmoji(emoji.skinVariations, emojiData.meta);
 
@@ -119,6 +119,7 @@ export const denormaliseEmojiServiceResponse = (emojiData: EmojiServiceResponse)
       representation,
       skinVariations,
       ascii,
+      searchable,
     };
   });