+++ packages/emoji/test/MockEmojiResource.ts
@@ -1,8 +1,8 @@
 import * as uid from 'uid';
 
 import { customCategory, customType } from '../src/constants';
-import { EmojiDescription, EmojiId, EmojiUpload, OptionalEmojiDescription, SearchOptions } from '../src/types';
-import { addCustomCategoryToResult, UploadingEmojiProvider } from '../src/api/EmojiResource';
+import { EmojiDescription, EmojiId, EmojiUpload, OptionalEmojiDescription, SearchOptions, ToneSelection } from '../src/types';
+import { addCustomCategoryToResult, EmojiProvider, UploadingEmojiProvider } from '../src/api/EmojiResource';
 import EmojiRepository, { EmojiSearchResult } from '../src/api/EmojiRepository';
 import { AbstractResource } from '../src/api/SharedResources';
 import debug from '../src/util/logger';
@@ -15,6 +15,7 @@ export interface MockEmojiResourceConfig {
   promiseBuilder?: PromiseBuilder<any>;
   uploadSupported?: boolean;
   uploadError?: string;
+  optimisticRendering?: boolean;
 }
 
 export const emojiFromUpload = (upload: EmojiUpload) => {
@@ -31,14 +32,17 @@ export const emojiFromUpload = (upload: EmojiUpload) => {
       width,
       height,
       imagePath: dataURL,
-    }
+    },
+    searchable: true
   };
 };
 
-export class MockNonUploadingEmojiResource extends AbstractResource<string, EmojiSearchResult, any, undefined, SearchOptions> {
+export class MockNonUploadingEmojiResource extends AbstractResource<string, EmojiSearchResult, any, undefined, SearchOptions> implements EmojiProvider {
   protected emojiRepository: EmojiRepository;
   protected promiseBuilder: PromiseBuilder<any>;
   protected lastQuery: string = '';
+  protected selectedTone: ToneSelection;
+  protected optimisticRendering?: boolean;
 
   recordedSelections: EmojiId[] = [];
 
@@ -50,6 +54,7 @@ export class MockNonUploadingEmojiResource extends AbstractResource<string, Emoj
       if (config.promiseBuilder) {
         this.promiseBuilder = config.promiseBuilder;
       }
+      this.optimisticRendering = config.optimisticRendering;
     }
   }
 
@@ -98,6 +103,23 @@ export class MockNonUploadingEmojiResource extends AbstractResource<string, Emoj
     this.recordedSelections.push(id);
     return this.promiseBuilder(undefined);
   }
+
+  loadMediaEmoji(emoji: EmojiDescription): OptionalEmojiDescription | Promise<OptionalEmojiDescription> {
+    return emoji;
+  }
+
+  optimisticMediaRendering(emoji: EmojiDescription) {
+    return !!this.optimisticRendering;
+  }
+
+  getSelectedTone(): ToneSelection {
+    return this.selectedTone;
+  }
+
+  setSelectedTone(tone: ToneSelection) {
+    this.selectedTone = tone;
+  }
+
 }
 
 export interface UploadDetail {
@@ -156,6 +178,12 @@ export class MockEmojiResource extends MockNonUploadingEmojiResource implements 
     super.notifyNotReady();
   }
 
+  loadMediaEmoji(emoji: EmojiDescription) {
+    if (this.promiseBuilder) {
+      return this.promiseBuilder(emoji);
+    }
+    return emoji;
+  }
 }
 
 export const mockNonUploadingEmojiResourceFactory = (emojiRepository: EmojiRepository, config?: MockEmojiResourceConfig, promiseBuilder?: PromiseBuilder<MockNonUploadingEmojiResource>) => {